+++ app/Http/Controllers/OrganizationController.php
@@ -82,8 +82,12 @@ class OrganizationController extends Controller
         return Redirect::route('organization.edit', $organizationId);
     }
 
-    public function delete($organizationId)
+    public function delete(Request $request, $organizationId)
     {
+        if ($request->user()->organization_id != $organizationId && !$request->user()->hasRole('super-admin-1')) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $organization = Organization::where('id', $organizationId)->first();
 
         $organization->delete();
